Protokoll zur Kommunikation teilweise implementiert - Simulatoren herausgelöst

+++ src/loggerbldcmotordriver/DiscretRandomSignalGenerator.java
@@ -1,5 +1,8 @@
 package loggerbldcmotordriver;
 
+import loggerbldcmotordriver.com.IntegerData;
+import loggerbldcmotordriver.com.LongData;
+
 /**
  *
  * @author simon
@@ -13,9 +16,15 @@ public class DiscretRandomSignalGenerator extends Thread
     private int dc_value;
 
     private final TimeReference timeReference;
-    private final RingBuffer<IntegerData> buffer;
+    private final RingBuffer<LongData> buffer;
 
-    public DiscretRandomSignalGenerator(int min_sleep_ms, int max_sleep_ms, int max_amplitude, int dc_value, TimeReference timeReference, RingBuffer<IntegerData> buffer) {
+    public DiscretRandomSignalGenerator(
+            int min_sleep_ms, 
+            int max_sleep_ms, 
+            int max_amplitude, 
+            int dc_value, 
+            TimeReference timeReference, 
+            RingBuffer<LongData> buffer) {
         this.min_sleep_ms = min_sleep_ms;
         this.max_sleep_ms = max_sleep_ms;
         this.max_amplitude = max_amplitude;
@@ -34,7 +43,7 @@ public class DiscretRandomSignalGenerator extends Thread
                 Thread.sleep(random_sleep_ms);
 
                 int random_value = (int) (Math.random() * (max_amplitude)) -max_amplitude/2 + dc_value;
-                buffer.put(new IntegerData(timeReference.getElapsedTime_ms(), random_value));
+                buffer.put(new LongData(timeReference.getElapsedTime_ms(), random_value));
             }
         }
         catch (Exception ex) {
