+++ src/loggerbldcmotordriver/IDataSink.java
@@ -1,11 +1,13 @@
 
 package loggerbldcmotordriver;
 
+import loggerbldcmotordriver.com.LongData;
+
 /**
  *
  * @author simon
  */
 public interface IDataSink
 {
-    public void put(IntegerData data);
+    public void put(LongData data);
 }
